+++ service/account/internal/src/test/resources/kapua-environment-setting.properties
@@ -61,3 +61,9 @@ commons.entity.insert.max.retry=3
 character.encoding=UTF-8
 
 commons.osgi.context=false
+
+# 
+# Entity settings
+# 
+commons.entity.key.size=8
+commons.entity.insert.max.retry=3
