+++ service/device/registry/internal/src/test/resources/kapua-environment-setting.properties
@@ -55,3 +55,9 @@ broker.port=1884
 character.encoding=UTF-8
 	
 commons.osgi.context=false
+
+# 
+# Entity settings
+# 
+commons.entity.key.size=8
+commons.entity.insert.max.retry=3
