+++ service/device/registry/internal/src/test/resources/locator.xml
@@ -19,16 +19,15 @@
 		<api>org.eclipse.kapua.service.account.AccountService</api>
 		<api>org.eclipse.kapua.service.user.UserFactory</api>
 		<api>org.eclipse.kapua.service.user.UserService</api>
-		<api>org.eclipse.kapua.service.generator.id.IdGeneratorService</api>
 		<api>org.eclipse.kapua.service.device.registry.DeviceRegistryService</api>
 		<api>org.eclipse.kapua.service.device.registry.event.DeviceEventService</api>
 	</provided>
 	<packages>
 		<package>org.eclipse.kapua.service.device.registry.internal</package>
 		<package>org.eclipse.kapua.service.device.registry.event.internal</package>
-		<package>org.eclipse.kapua.service.generator.id.sequence</package>
 		<package>org.eclipse.kapua.test.user</package>
 		<package>org.eclipse.kapua.test.account</package>
-		<package>org.eclipse.kapua.test.auth</package>
+		<package>org.eclipse.kapua.test.authentication</package>
+		<package>org.eclipse.kapua.test.authorization</package>
 	</packages>
 </locator-config>
