+++ service/security/shiro/src/main/java/org/eclipse/kapua/service/authorization/access/shiro/AccessInfoServiceImpl.java
@@ -18,6 +18,7 @@ import org.eclipse.kapua.commons.jpa.AbstractEntityManagerFactory;
 import org.eclipse.kapua.commons.service.internal.AbstractKapuaService;
 import org.eclipse.kapua.commons.util.ArgumentValidator;
 import org.eclipse.kapua.locator.KapuaLocator;
+import org.eclipse.kapua.locator.KapuaProvider;
 import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.model.query.KapuaQuery;
 import org.eclipse.kapua.service.authorization.AuthorizationService;
@@ -49,6 +50,7 @@ import org.eclipse.kapua.service.authorization.shiro.KapuaAuthorizationException
  * 
  * @since 1.0.0
  */
+@KapuaProvider
 public class AccessInfoServiceImpl extends AbstractKapuaService implements AccessInfoService {
 
     /**
