+++ test/src/main/java/org/eclipse/kapua/test/authentication/AuthenticationServiceMock.java
@@ -10,7 +10,7 @@
  *     Eurotech - initial API and implementation
  *
  *******************************************************************************/
-package org.eclipse.kapua.test.auth;
+package org.eclipse.kapua.test.authentication;
 
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.commons.security.KapuaSecurityUtils;
